Move allow_cors to a module instead of monkey patch
