--- lib/davinci_dtr_test_kit/ext/inferno_core/runnable.rb
@@ -1,22 +1,0 @@
-module Inferno
-  module DSL
-    module Runnable
-      PRE_FLIGHT_HANDLER = proc do
-        [
-          200,
-          {
-            'Access-Control-Allow-Origin' => '*',
-            'Access-Control-Allow-Headers' => 'Content-Type, Authorization'
-          },
-          ['']
-        ]
-      end
-
-      def allow_cors(*paths)
-        paths.each do |path|
-          route(:options, path, PRE_FLIGHT_HANDLER)
-        end
-      end
-    end
-  end
-end
